pyproject.toml
[project]
name = "multi-agent-sales-development-representative-automation-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.8.0",
    "openai-agents>=0.5.1",
    "python-dotenv>=1.2.1",
    "sendgrid>=6.12.5",
]
